refactor(GetResult) : 당첨 결과를 판정하는 인터페이스를 구현하는 방식으로 변경
- LottoMatcher -> LottoREsultEvaluator 로 변경
