+++ src/main/java/lotto/Application.java
@@ -1,7 +1,7 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.domain.LottoMatcher;
+import lotto.domain.lottoresultevaluator.LottoResultEvaluatorImpl;
 import lotto.domain.lottogenerator.RandomLottoGenerator;
 import lotto.service.LottoPurchaseService;
 import lotto.service.LottoResultService;
@@ -14,7 +14,7 @@ public class Application {
                 new InputView(),
                 new OutputView(),
                 new LottoPurchaseService(new RandomLottoGenerator()),
-                new LottoResultService(new LottoMatcher()));
+                new LottoResultService(new LottoResultEvaluatorImpl()));
         lottoController.run();
     }
 }
