+++ src/main/java/lotto/domain/lottoresultevaluator/LottoResultEvaluatorImpl.java
@@ -1,17 +1,20 @@
-package lotto.domain;
+package lotto.domain.lottoresultevaluator;
 
 import lotto.config.WinningRank;
+import lotto.domain.Lotto;
+import lotto.domain.WinningNumbers;
 
 import java.util.*;
 
-public class LottoMatcher {
+public class LottoResultEvaluatorImpl implements LottoResultEvaluator {
     private int match(List<Integer> lotto, List<Integer> mainNumber) {
         Set<Integer> intersection = new HashSet<>(lotto);
         intersection.retainAll(mainNumber);
         return intersection.size();
     }
 
-    public WinningRank getResult(Lotto lotto, WinningNumbers winningNumbers) {
+    @Override
+    public WinningRank evaluate(Lotto lotto, WinningNumbers winningNumbers) {
         int matchCount = match(lotto.getNumbers(), winningNumbers.getMainNumbers().getNumbers());
         boolean bonusMatched = lotto.getNumbers().contains(winningNumbers.getBonusNumber().getValue());
         return WinningRank.of(matchCount, bonusMatched);
